# Factor performance - `vec_cast()`