--- a/src/cast-dispatch.c
+++ b/src/cast-dispatch.c
@@ -1,0 +1,90 @@
+#include "vctrs.h"
+#include "utils.h"
+
+static SEXP vec_cast_dispatch2(SEXP x,
+                               SEXP to,
+                               enum vctrs_type x_type,
+                               bool* lossy,
+                               struct vctrs_arg* x_arg,
+                               struct vctrs_arg* to_arg);
+
+// [[ include("vctrs.h") ]]
+SEXP vec_cast_dispatch(SEXP x,
+                       SEXP to,
+                       enum vctrs_type x_type,
+                       enum vctrs_type to_type,
+                       bool* lossy,
+                       struct vctrs_arg* x_arg,
+                       struct vctrs_arg* to_arg) {
+  switch (to_type) {
+  case vctrs_type_character:
+    switch (class_type(x)) {
+    case vctrs_class_bare_factor:
+      return fct_as_character(x, x_arg);
+
+    case vctrs_class_bare_ordered:
+      return ord_as_character(x, x_arg);
+
+    default:
+      break;
+    }
+
+  case vctrs_type_s3:
+    return vec_cast_dispatch2(x, to, x_type, lossy, x_arg, to_arg);
+
+  default:
+    break;
+  }
+
+  return R_NilValue;
+}
+
+
+static SEXP vec_cast_dispatch2(SEXP x, SEXP to,
+                               enum vctrs_type x_type,
+                               bool* lossy,
+                               struct vctrs_arg* x_arg,
+                               struct vctrs_arg* to_arg) {
+  switch (class_type(to)) {
+  case vctrs_class_bare_factor:
+    switch (x_type) {
+    case vctrs_type_character:
+      return chr_as_factor(x, to, lossy, to_arg);
+
+    case vctrs_type_s3:
+      switch (class_type(x)) {
+      case vctrs_class_bare_factor:
+        return fct_as_factor(x, to, lossy, x_arg, to_arg);
+
+      default:
+        break;
+      }
+
+    default:
+      break;
+    }
+
+  case vctrs_class_bare_ordered:
+    switch (x_type) {
+    case vctrs_type_character:
+      return chr_as_ordered(x, to, lossy, to_arg);
+
+    case vctrs_type_s3:
+      switch (class_type(x)) {
+      case vctrs_class_bare_ordered:
+        return ord_as_ordered(x, to, lossy, x_arg, to_arg);
+
+      default:
+        break;
+      }
+
+    default:
+      break;
+    }
+
+  default:
+    break;
+  }
+
+  return R_NilValue;
+}