--- a/src/cast.c
+++ b/src/cast.c
@@ -267,9 +267,16 @@ static SEXP df_as_dataframe(SEXP x, SEXP to, struct vctrs_arg* x_arg, struct vct
 }
 
 static SEXP vec_cast_switch(SEXP x, SEXP to, bool* lossy, struct vctrs_arg* x_arg, struct vctrs_arg* to_arg) {
-  switch (vec_typeof(to)) {
+  enum vctrs_type to_type = vec_typeof(to);
+  enum vctrs_type x_type = vec_typeof(x);
+
+  if (to_type == vctrs_type_s3 || x_type == vctrs_type_s3) {
+    return vec_cast_dispatch(x, to, x_type, to_type, lossy, x_arg, to_arg);
+  }
+
+  switch (to_type) {
   case vctrs_type_logical:
-    switch (vec_typeof(x)) {
+    switch (x_type) {
     case vctrs_type_logical:
       return x;
     case vctrs_type_integer:
@@ -284,7 +291,7 @@ static SEXP vec_cast_switch(SEXP x, SEXP to, bool* lossy, struct vctrs_arg* x_ar
     break;
 
   case vctrs_type_integer:
-    switch (vec_typeof(x)) {
+    switch (x_type) {
     case vctrs_type_logical:
       return lgl_as_integer(x, lossy);
     case vctrs_type_integer:
@@ -300,7 +307,7 @@ static SEXP vec_cast_switch(SEXP x, SEXP to, bool* lossy, struct vctrs_arg* x_ar
     break;
 
   case vctrs_type_double:
-    switch (vec_typeof(x)) {
+    switch (x_type) {
     case vctrs_type_logical:
       return lgl_as_double(x, lossy);
     case vctrs_type_integer:
@@ -316,7 +323,7 @@ static SEXP vec_cast_switch(SEXP x, SEXP to, bool* lossy, struct vctrs_arg* x_ar
     break;
 
   case vctrs_type_character:
-    switch (vec_typeof(x)) {
+    switch (x_type) {
     case vctrs_type_logical:
     case vctrs_type_integer:
     case vctrs_type_double:
@@ -329,7 +336,7 @@ static SEXP vec_cast_switch(SEXP x, SEXP to, bool* lossy, struct vctrs_arg* x_ar
     break;
 
   case vctrs_type_dataframe:
-    switch (vec_typeof(x)) {
+    switch (x_type) {
     case vctrs_type_dataframe:
       return df_as_dataframe(x, to, x_arg, to_arg);
     default: