--- a/src/init.c
+++ b/src/init.c
@@ -30,7 +30,7 @@ extern SEXP vctrs_group_id(SEXP);
 extern SEXP vctrs_group_rle(SEXP);
 extern SEXP vec_group_loc(SEXP);
 extern SEXP vctrs_equal(SEXP, SEXP, SEXP);
-extern SEXP vctrs_equal_na(SEXP);
+extern SEXP vec_equal_na(SEXP);
 extern SEXP vctrs_compare(SEXP, SEXP, SEXP);
 extern SEXP vec_match(SEXP, SEXP);
 extern SEXP vctrs_duplicated_any(SEXP);
@@ -135,7 +135,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_dim_n",                      (DL_FUNC) &vctrs_dim_n, 1},
   {"vctrs_is_unspecified",             (DL_FUNC) &vctrs_is_unspecified, 1},
   {"vctrs_equal",                      (DL_FUNC) &vctrs_equal, 3},
-  {"vctrs_equal_na",                   (DL_FUNC) &vctrs_equal_na, 1},
+  {"vctrs_equal_na",                   (DL_FUNC) &vec_equal_na, 1},
   {"vctrs_compare",                    (DL_FUNC) &vctrs_compare, 3},
   {"vctrs_match",                      (DL_FUNC) &vec_match, 2},
   {"vctrs_typeof",                     (DL_FUNC) &vctrs_typeof, 2},