… src/type2.c
… src/vctrs.h
… tests/testthat/test-type-factor.R