This PR adds internal `vec_cast()` dispatch for factors/ordered factors.

It does so in a way that is similar to `vec_ptype2()`, using the extension prototype idea @lionel- and I used there. I've added a `vec_cast_dispatch()` path in `vec_cast_switch()`, which allows us to hand off to specific C level casting routines for factors. The routines themselves live in `type-factor.c`.

I've maintained all the behavior from before, including the "template" idea of doing `vec_cast(factor("x"), factor())`.

A number of performance improvements were made, the largest being the easy case where `x` and `to` are both factors with the same levels, in which case there is nothing to do.

```r
library(vctrs)

vec_cast.factor.factor <- vctrs:::vec_cast.factor.factor
vec_cast.factor.character <- vctrs:::vec_cast.factor.character

a <- factor(rep("a", 1e5))
```

Casting to a factor with the same levels is now very fast

```r
bench::mark(
  vec_cast(a, a),
  vec_cast.factor.factor(a, a)
)
#> # A tibble: 2 x 6
#>   expression                        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(a, a)                  897ns   1.21µs   790069.        0B     79.0
#> 2 vec_cast.factor.factor(a, a)   2.98ms   3.84ms      242.    4.61MB     24.2
```

Same with casting to a template

```r
template <- factor()

bench::mark(
  vec_cast(a, template),
  vec_cast.factor.factor(a, template)
)
#> # A tibble: 2 x 6
#>   expression                               min median `itr/sec` mem_alloc
#>   <bch:expr>                          <bch:tm> <bch:>     <dbl> <bch:byt>
#> 1 vec_cast(a, template)                  951ns 1.21µs   761576.        0B
#> 2 vec_cast.factor.factor(a, template)   1.16ms 1.54ms      637.    1.91MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

The factor to factor conversion tries to be somewhat smart when the levels are different, and the levels of `x` are a subset of the levels of `to`. If the levels of `x` are a contiguous subset of the levels of `to` then we can just construct a new factor immediately without "recoding" the underlying integer values.

```r
# no recoding required because levels of `a` are a contiguous subset of `ab`
# i.e. c("a") is a subset of c("a", "b") and is in the same order
ab <- factor(c(rep("a", 1e5/2), rep("b", 1e5/2)), levels = c("a", "b"))

bench::mark(
  vec_cast(a, ab),
  vec_cast.factor.factor(a, ab),
  factor(a, levels = levels(ab))
)
#> # A tibble: 3 x 6
#>   expression                          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(a, ab)                 31.03µs  57.17µs    14451.  390.67KB    122. 
#> 2 vec_cast.factor.factor(a, ab)    2.87ms   3.55ms      284.    4.58MB     29.4
#> 3 factor(a, levels = levels(ab))   1.36ms   1.48ms      672.    1.91MB     34.6
```

Here is a small comparison of recoding vs not recoding

```r
# recoding required because levels of `a` are NOT a contiguous subset of `ba`
# i.e. c("a") is a subset of c("b", "a") but is NOT in order
ba <- factor(c(rep("b", 1e5/2), rep("a", 1e5/2)), levels = c("b", "a"))

bench::mark(
  vec_cast(a, ab),
  vec_cast(a, ba),
  check = FALSE,
  iterations = 5000
)
#> # A tibble: 2 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(a, ab)   30.2µs   55.5µs    15426.     391KB    118. 
#> 2 vec_cast(a, ba)   53.7µs   81.2µs    12208.     391KB     86.1
```

Here is an example of what is probably the worst case scenario. Every value is a different level. Faster than the original implementation still.

```r
# probably the slowest it would ever be
many_levels <- factor(paste0("a", 1:1e5), levels = paste0("a", 1:1e5))
many_levels_rev <- factor(rev(paste0("a", 1:1e5)), levels = rev(paste0("a", 1:1e5)))

bench::mark(
  vec_cast(many_levels, many_levels_rev),
  vec_cast.factor.factor(many_levels, many_levels_rev),
  factor(many_levels, levels = levels(many_levels_rev)),
  iterations = 50
)
#> # A tibble: 3 x 6
#>   expression                                               min median `itr/sec`
#>   <bch:expr>                                            <bch:> <bch:>     <dbl>
#> 1 vec_cast(many_levels, many_levels_rev)                10.6ms 12.2ms      80.7
#> 2 vec_cast.factor.factor(many_levels, many_levels_rev)  18.1ms 19.4ms      51.0
#> 3 factor(many_levels, levels = levels(many_levels_rev)) 12.4ms   13ms      76.3
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>
```

I've also added character -> factor C routines. The one where `to` is a template takes the longest because we have to take the unique values of `x` to get the levels, run through them and remove the `NA` level if it exists, and then match that against `x`.

```r
chr <- paste0("a", 1:3e5)
fct <- factor(chr, levels = chr)

bench::mark(
  vec_cast(chr, fct),
  vec_cast.factor.character(chr, fct),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                          <bch:> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_cast(chr, fct)                  36.9ms 39.6ms      23.8    5.43MB     5.71
#> 2 vec_cast.factor.character(chr, fct) 74.2ms   85ms      11.3   34.89MB    14.7

bench::mark(
  vec_cast(chr, template),
  vec_cast.factor.character(chr, template),
  iterations = 50
)
#> # A tibble: 2 x 6
#>   expression                                  min median `itr/sec` mem_alloc
#>   <bch:expr>                               <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_cast(chr, template)                  56.5ms 60.4ms      16.6      16MB
#> 2 vec_cast.factor.character(chr, template) 66.7ms 66.7ms      15.0    30.3MB
#> # … with 1 more variable: `gc/sec` <dbl>
```