@DavisVaughan 2020-02-11
> Would it be much work to handle the lossy cases using the new lazy error infrastructure, instead of falling back to the R methods?

Together we decided to leave this for a separate PR