@DavisVaughan 2020-02-11
Improvement on `vec_c()`-ing many factors with the same levels. This is related to one of the dplyr issues regarding slowness b/c of vctrs

```r
library(vctrs)

x <- factor(letters)
x <- rep_len(list(x), 10000)
head(x, 2)
#> [[1]]
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z
#> Levels: a b c d e f g h i j k l m n o p q r s t u v w x y z
#> 
#> [[2]]
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z
#> Levels: a b c d e f g h i j k l m n o p q r s t u v w x y z

# Before
bench::mark(vec_c(!!!x))
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_c(!!!x)    647ms    647ms      1.55    43.6MB     20.1

# After
bench::mark(vec_c(!!!x))
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_c(!!!x)   24.3ms   26.2ms      38.2    3.12MB     78.3
```